# Deprecation warning newer pymongo #1202