--- a/setup.py
+++ b/setup.py
@@ -15,7 +15,7 @@ INSTALL_REQUIRES = [
     "cerberus>=1.1",
     "events>=0.3,<0.4",
     "flask>=1.0",
-    "pymongo>=3.5",
+    "pymongo>=3.7",
     "simplejson>=3.3.0,<4.0",
 ]