PR to get some comments on the implementation ref discussion for #1202.  No tests yet.

Main takeaways are:

- needs pymongo>=3.7
- breaks datalayer `find()` return signature from `cursor` to `count, cursor`
- introduces new global config setting `PAGINATION_STRATEGY` and per resource `pagination_strategy`

 `PAGINATION_STRATEGY` will probably make `OPTIMIZE_PAGINATION_FOR_SPEED` obsolete. Options are

- `None` no count
- `full` pymongos `count_documents` which do an aggregation query over all documents hence is accurate, equivalent to deprecated `cursor.count()`
- `estimated` pymongos `estimated_document_count` which estimates the number of documents based on collection statistics

I haven't tried  `count_documents` hint and max execution time options, could possibly be something to investigate?

I did some simple benchmarks with wrk. On the smaller collections there is not much difference, but it's appearent that `count_documents` becomes increasingly slower with collection size.

On this PR,  `PAGINATION_STRATEGY`:

- `None` 216 requests in 20.02s
- `estimated` 191 requests in 20.03s
- `full` 9 requests in 20.04s

On pyeve/master,  `OPTIMIZE_PAGINATION_FOR_SPEED`:

- `False` 9 requests in 20.03s
- `True` 206 requests in 20.03s

Tests done on an indexed collection just shy of 5m documents, one connection, to be taken as an indication only.

I did observe `estimated` to be accurate on all requests. Also, the very first database hit for `estimated` takes around 1/3 of the time of an `full` then drops by a factor of 10 and is constant for all following requests. I assume that `estimated` will be accurate and almost as fast as no counting for large collections with few updates. With frequent updates it would probably become slower but still a lot faster than doing the `full` count. I don't know about accuracy on frequent updates.

I think this could be of great use especially for those with large collections.

Feedback appreciated!