@einarhuseby 2019-01-29
Something is not right with pagination, and after some RTFM it's clear to me that `estimated_document_count` *only* gets the total number of documents - and not by filter.

However,  `estimated_document_count` should still be useful when no filter/counting all documents in a collection.

Also need to look into the replacement of query operators for [count_documents](http://api.mongodb.com/python/current/api/pymongo/collection.html#pymongo.collection.Collection.count_documents)