@nicolaiarocci 2019-03-26
Alright, I took a slightly different approach (see https://github.com/pyeve/eve/commit/15f5e650724b2a7503312f739f5442e5e2269e20), thank you for showing me the path.

Also, I focused on the deprecation warning. Feel free to open a new PR with work on pagination optimization.