@nicolaiarocci 2019-03-27
Amended to only run `count_documents()` when required, see https://github.com/pyeve/eve/commit/ac3bd78a9dfcad38e2fe478c0f35f66ac290c37c.